modify toy-isa reg_t and op_t as int && modify EBPF_FLAG as ISA_EBPF

+++ src/isa/toy-isa/inst.cc
@@ -41,9 +41,9 @@ toy_isa_inst& toy_isa_inst::operator=(const inst &rhs) {
 }
 
 // For jmp opcode, it can only jump forward
-int32_t toy_isa_inst::get_max_operand_val(int op_index, int inst_index) const {
+int toy_isa_inst::get_max_operand_val(int op_index, int inst_index) const {
   // max value for each operand type
-  int32_t max_val[4] = {
+  int max_val[4] = {
     [toy_isa::OP_UNUSED] = 0,
     [toy_isa::OP_REG] = toy_isa::NUM_REGS,
     [toy_isa::OP_IMM] = toy_isa::MAX_CONST,
@@ -129,7 +129,7 @@ void toy_isa_inst::set_as_nop_inst() {
   _args[2] = 0;
 }
 
-int64_t toy_isa_inst::interpret(const vector<inst*> &instptr_list, prog_state &ps, int64_t input) const {
+int toy_isa_inst::interpret(const vector<inst*> &instptr_list, prog_state &ps, int input) const {
   /* Input currently is just one integer which will be written into R0. Will
   need to generalize this later. */
   // inst *insn = this;
